+++ src/rics/envinterp/__init__.py
@@ -1,7 +1,6 @@
 """Legacy module.
 
 Schedule:
-    * ``rics==0.6.0``: Emit :py:class:`DeprecationWarning` on import.
     * ``rics==0.7.0``: Drop legacy module.
 
 Please use :mod:`rics.env.interpolation` instead.
@@ -12,11 +11,19 @@ from rics.env.interpolation import UnsetVariableError, Variable, replace_in_stri
 __all__ = ["UnsetVariableError", "Variable", "replace_in_string"]
 
 import os
+import warnings
 
 if os.environ.get("SPHINX_BUILD") == "true":
     # Ugly fix for duplicate indexes.
     del __all__, UnsetVariableError, Variable, replace_in_string
+
+warnings.warn(
+    f"Package `{__package__}` is deprecated. Use `rics.env.interpolation` instead.",
+    UserWarning,
+    stacklevel=2,
+)
+
+del warnings
 del os
 
-# TODO(6.0.0): Deprecate this module
 # TODO(7.0.0): Remove this module