+++ src/rics/misc.py
@@ -18,9 +18,16 @@ def interpolate_environment_variables(
     allow_blank: bool = False,
 ) -> str:
     """Alias of :func:`rics.env.interpolation.replace_in_string`."""
+    import warnings
+
     from rics.env.interpolation import replace_in_string
 
-    # TODO(6.0.0): Deprecate this function
+    warnings.warn(
+        "Function `interpolate_environment_variables()` is deprecated."
+        "Use `rics.env.interpolation.replace_in_string` instead.",
+        UserWarning,
+        stacklevel=2,
+    )
     # TODO(7.0.0): Remove this function
 
     return replace_in_string(s, allow_nested=allow_nested, allow_blank=allow_blank)