+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Make `pandas` an optional dependency.
 
+### Deprecated
+- The `rics.envinterp` module
+- The `mics.interpolate_environment_variables()` function
+
 ## [5.1.2] - 2025-06-14
 
 ### Fixed